Fixed issue with refresh, removed function from reducer and placed in action file
