+++ frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "frontend",
-  "version": "1.0.2",
+  "version": "1.0.3",
   "private": true,
   "homepage": "http://localhost:8080",
   "dependencies": {
