+++ frontend/src/components/cluster-dashboard/ClusterDashboardActions.tsx
@@ -2,10 +2,11 @@
 import { Dispatch } from "react";
 import axios from "axios";
 import querystring from "query-string";
-
+import React from 'react';
+import ReactDOMServer from 'react-dom/server';
+import GraphNodeTooltip from './graph-node-tooltip/GraphNodeTooltip';
 import { Cluster } from "./Cluster.model";
 import { ClusterDashboardState, initialState } from "./ClusterDashboardReducer";
-
 import stripJsonComments from "strip-json-comments";
 
 export const FETCH_CLUSTER_MEMBERS_BEGIN = 'FETCH_CLUSTER_MEMBERS_BEGIN';
@@ -47,7 +48,7 @@ interface ChangeRefreshIntervalAction {
 
 interface FrameGraphDataAction {
   type: typeof FRAME_GRAPH_DATA;
-  payload: { styles: any; nodeUrl: string; clusterUrl: string };
+  payload: { graph: {nodes: any[], edges: any[]} };
 }
 
 interface AddClusterNodeBeginAction {
@@ -111,7 +112,6 @@ export const fetchClusterMembersBegin = (): FetchClusterMembersBeginAction => ({
 });
 
 export const fetchClusterMembersSuccess = (cluster: Cluster): FetchClusterMembersSuccessAction => {
-	console.log("TCL: cluster", cluster)
   return({
   type: FETCH_CLUSTER_MEMBERS_SUCCESS,
   payload: { cluster: cluster }
@@ -128,10 +128,10 @@ export const changeRefreshInterval =
     payload: { state: state }
   });
 
-export const frameGraphData = (styles: any, nodeUrl: string, clusterUrl: string):
+export const frameGraphData = (graph: {nodes: any[], edges: any[]}):
   FrameGraphDataAction => ({
     type: FRAME_GRAPH_DATA,
-    payload: { styles: styles, nodeUrl: nodeUrl, clusterUrl: clusterUrl }
+    payload: { graph },
   });
 
 export const addClusterNodeBegin = (): AddClusterNodeBeginAction => ({
@@ -200,7 +200,10 @@ export function fetchClusterData() {
     const akkaManagementUrl = getState().dashboard.akkaProps.managementUrl;
     return axios
       .get(`${akkaManagementUrl}/cluster/members`)
-      .then(response => dispatch(fetchClusterMembersSuccess(response.data)))
+      .then(response => {
+				console.log("TCL: fetchClusterData -> response ", response)
+        dispatch(fetchClusterMembersSuccess(response.data))
+      })
       .catch(error => dispatch(fetchClusterMembersFailure(error)));
   };
 }
@@ -261,6 +264,92 @@ export function fetchAkkaProps() {
       })
       .catch(error => dispatch(fetchAkkaPropsFailure(error)));
   };
+}
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const setupGraph = 
+(cluster: Cluster, styles: any, nodeUrl: string, clusterUrl: string): any => 
+(dispatch: any) => {
+  const leader = {
+    shadow: {
+      enabled: true,
+      color: styles.leaderNodeColor,
+      size: 15,
+      x: 1,
+      y: 1
+    }
+  };
+
+  const oldest = {
+    borderWidth: 3,
+    borderWidthSelected: 0,
+    color: {
+      border: styles.oldestNodeColor,
+      highlight: { border: styles.oldestNodeColor },
+      hover: { border: styles.oldestNodeColor }
+    },
+    shapeProperties: { borderDashes: [10, 10] }
+  };
+
+  //eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const graph: any = {
+    nodes: [
+      {
+        id: 0,
+        label: "Akka Cluster",
+        image: clusterUrl,
+        size: 40,
+        borderWidth: 1,
+        borderWidthSelected: 2,
+        color: {
+          border: styles.primaryColor,
+          background: styles.secondaryColor,
+          highlight: {
+            border: styles.primaryColor,
+            background: styles.secondaryColorLighter
+          },
+          hover: {
+            border: styles.primaryColor,
+            background: styles.secondaryColorLighter
+          }
+        }
+      }
+    ],
+    edges: []
+  };
+
+  if (!cluster || !cluster.members) {
+    return graph;
+  }
+
+  cluster.members.forEach((member, index) => {
+    const memberTitle = <GraphNodeTooltip member={member} clusterData={cluster} />;
+
+    const memberConfig = {
+      id: index + 1,
+      label: `<b>o </b>${member.node.split("://")[1]}`,
+      image: nodeUrl,
+      title: ReactDOMServer.renderToString(memberTitle),
+      font: {
+        bold: {
+          color: styles[`status${member.status}Color`],
+          size: 16,
+          vadjust: -0.5
+        },
+        multi: true
+      },
+      ...(member.node === cluster.leader && leader),
+      ...(member.node === cluster.oldest && oldest)
+    };
+
+    
+		console.log("TCL: memberconfig", memberConfig, cluster);
+
+    graph.nodes.push(memberConfig);
+    graph.edges.push({ from: 0, to: index + 1 });
+  });
+
+  dispatch(frameGraphData(graph));
 }
 
 export type ClusterDashboardActionTypes = FetchClusterMembersBeginAction |
